encoding: delete v1 proto codec and prefer v1 codecs over v2 codecs
