+++ encoding/proto/proto_test.go
@@ -25,10 +25,11 @@ import (
 
 	"google.golang.org/grpc/encoding"
 	"google.golang.org/grpc/internal/grpctest"
+	"google.golang.org/grpc/mem"
 	pb "google.golang.org/grpc/test/codec_perf"
 )
 
-func marshalAndUnmarshal(t *testing.T, codec encoding.Codec, expectedBody []byte) {
+func marshalAndUnmarshal(t *testing.T, codec encoding.CodecV2, expectedBody []byte) {
 	p := &pb.Buffer{}
 	p.Body = expectedBody
 
@@ -55,7 +56,7 @@ func Test(t *testing.T) {
 }
 
 func (s) TestBasicProtoCodecMarshalAndUnmarshal(t *testing.T) {
-	marshalAndUnmarshal(t, codec{}, []byte{1, 2, 3})
+	marshalAndUnmarshal(t, &codecV2{}, []byte{1, 2, 3})
 }
 
 // Try to catch possible race conditions around use of pools
@@ -75,7 +76,7 @@ func (s) TestConcurrentUsage(t *testing.T) {
 	}
 
 	var wg sync.WaitGroup
-	codec := codec{}
+	codec := &codecV2{}
 
 	for i := 0; i < numGoRoutines; i++ {
 		wg.Add(1)
@@ -93,8 +94,8 @@ func (s) TestConcurrentUsage(t *testing.T) {
 // TestStaggeredMarshalAndUnmarshalUsingSamePool tries to catch potential errors in which slices get
 // stomped on during reuse of a proto.Buffer.
 func (s) TestStaggeredMarshalAndUnmarshalUsingSamePool(t *testing.T) {
-	codec1 := codec{}
-	codec2 := codec{}
+	codec1 := &codecV2{}
+	codec2 := &codecV2{}
 
 	expectedBody1 := []byte{1, 2, 3}
 	expectedBody2 := []byte{4, 5, 6}
@@ -102,7 +103,7 @@ func (s) TestStaggeredMarshalAndUnmarshalUsingSamePool(t *testing.T) {
 	proto1 := pb.Buffer{Body: expectedBody1}
 	proto2 := pb.Buffer{Body: expectedBody2}
 
-	var m1, m2 []byte
+	var m1, m2 mem.BufferSlice
 	var err error
 
 	if m1, err = codec1.Marshal(&proto1); err != nil {
